[BACKPORT 2.14][#13295] YCQL: Disable PAGING & LIMIT for aggregate functions
Summary:
Currently YCQL SELECT statement supports 5 aggregate functions: `count`, `avg`, `min`, `max`, `sum`.
These functions must return just a single row of result to the user. Paging does not make sense for such
aggregated SELECTs.

However, before the fix, an aggregate SELECT statement sends a single aggregated result back to the user
for every `P` tablets (`P` is the page size) along with a paging state which prompts the user to query again and
obtain another aggregate on the next `P` tablets and so on. This occurs as follows:

(1) Assume page size is set to `P` on the ycqlsh prompt.
(2) A tablet `test` has with data spread across `N` tablets. Assume that no tablet is empty for ease of description.
(3) An aggregate SELECT is issued by the user and the YCQL proxy uses a `max_fetch_size` of `P` for it.
This means that the YCQL proxy will keep querying aggregates from tablets till it gets `P` aggregate values.
(4) The YCQL proxy scans tablets one-by-one sequentially via a `ReadRequestPB` RPC. On receiving this RPC,
the TServer scans all rows in the tablet to return 1 row to the YCQL proxy. Each tablet returns 1 aggregate row.
The YCQL proxy then scans P-1 more tablets until it gets P aggregates in total or no more tablets exist.
(5) Once P rows are collected or no more tablets exist for scanning, 1 single aggregate over P tablets is sent back
to the user. If more tablets exist for scanning, a paging state is sent back to the user for indicating that there is
still a part of the table that hasn't been scanned.

With this, an aggregate will return `ceiling(N/P)` rows.

E.g.:
```
create keyspace temp;
use temp;
create table test (k int primary key, v int) with tablets = 4 AND transactions = { 'enabled' : true };
insert into test (k, v) values (1, 1);
insert into test (k, v) values (2, 2);
insert into test (k, v) values (3, 3);
insert into test (k, v) values (4, 4);
insert into test (k, v) values (5, 5);
insert into test (k, v) values (6, 6);
paging 2;
select min(v) from test;
 system.min(v)
---------------
             1
---MORE---
 system.min(v)
---------------
             2
(2 rows)
```

Step (5) happens because iterating via all tablets is stopped in `Executor`
where `fetch_limit` (equals to the `PageSize`=5000 by default) is reached:
```
Result<bool> Executor::FetchMoreRows(const PTSelectStmt* tnode, ...) {
...
  int64_t fetch_limit = query_state->max_fetch_size();
  if (fetch_limit >= 0 && current_fetch_row_count >= fetch_limit) {
...
    return false;      <<<
```

The fix is to ensure that the YCQL proxy scans the full table for computing the aggregate and
doesn't stop after scanning `P` tablets. This can be done via ignoring the `max_fetch_size`
value for the aggregate functions in the top level `SELECT` statement.

`LIMIT` value is also ignored for the aggregate functions. E.g.:
```
select min(v) from test LIMIT 1;
```
Processing of `OFFSET` for the aggregate functions is NOT changed. E.g.:
```
select count(v) from test OFFSET 1;
```
Usage of `OFFSET` is a way to run the aggregate function for a random subset of rows
(skipping first N tablets).

Original commit: ce5f3da77518edb329148c62ca103847b2d178eb / D21444
GH link: https://github.com/yugabyte/yugabyte-db/commit/ce5f3da77518edb329148c62ca103847b2d178eb

Test Plan:
ybd --cxx-test cql-test --gtest_filter CqlTest.SelectAggregateFunctions
ybd --cxx-test ql_ql-query-test --gtest_filter TestQLQuery.TestPagingState

ybd --java-test org.yb.cql.TestSelect#testSelectWithOffset
ybd --java-test org.yb.cql.TestSelectRowCounter#testRowCounterSecondaryScan
ybd --java-test org.yb.cql.TestSelectRowCounter#testRowCounterPrimaryScan
ybd --java-test org.yb.cql.TestSelectRowCounter#testRowCounterPaging
ybd --java-test org.yb.cql.TestBindVariable#testBindingLimit

Reviewers: mihnea, neil, arybochkin, pjain

Reviewed By: pjain

Subscribers: yql

Differential Revision: https://phabricator.dev.yugabyte.com/D22212

+++ src/yb/yql/cql/ql/exec/exec_context.cc
@@ -452,7 +452,7 @@ QueryPagingState::QueryPagingState(const StatementParameters& user_params,
 
   // Just default it to max_int.
   if (max_fetch_size_ <= 0) {
-    max_fetch_size_ = INT_MAX;
+    max_fetch_size_ = INT64_MAX;
   }
 }
